backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
arxiv>=2.0.0
apscheduler>=3.10.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
cryptography>=42.0.0
openai>=1.30.0
python-multipart>=0.0.9
